+++ src/main/java/shop/geeksasang/GeeksasangApplication.java
@@ -4,6 +4,11 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+
 @EnableJpaAuditing
 @SpringBootApplication
 public class GeeksasangApplication {
@@ -12,4 +17,13 @@ public class GeeksasangApplication {
 		SpringApplication.run(GeeksasangApplication.class, args);
 	}
 
+	@Bean
+	public WebMvcConfigurer corsConfigurer() {
+		return new WebMvcConfigurer() {
+			@Override
+			public void addCorsMappings(CorsRegistry registry) {
+				registry.addMapping("/**").allowedOrigins("https://localhost:8080/**");
+			}
+		};
+	}
 }
