--- src/main/java/shop/geeksasang/advice/BaseExceptionAdvice.java
@@ -1,18 +1,0 @@
-package shop.geeksasang.advice;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import shop.geeksasang.config.exception.BaseException;
-import shop.geeksasang.config.exception.BaseResponseStatus;
-import shop.geeksasang.config.response.BaseResponse;
-
-@Slf4j
-@RestControllerAdvice
-public class BaseExceptionAdvice {
-    @ExceptionHandler(BaseException.class)
-    public BaseResponse<BaseResponseStatus> baseException(BaseException e) {
-        log.error("Handle CommonException: {}", e.getMessage());
-        return new BaseResponse<>(e.getStatus());
-    }
-}
