+++ src/main/java/shop/geeksasang/advice/GlobalExceptionAdvice.java
@@ -1,21 +1,66 @@
 package shop.geeksasang.advice;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.Order;
+import org.springframework.context.MessageSource;
 import org.springframework.http.HttpStatus;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
+import shop.geeksasang.config.exception.BaseException;
 import shop.geeksasang.config.exception.BaseResponseStatus;
 import shop.geeksasang.config.response.BaseResponse;
 
+import java.io.UnsupportedEncodingException;
+import java.security.InvalidKeyException;
+import java.security.NoSuchAlgorithmException;
+
 @Slf4j
-@Order(Ordered.LOWEST_PRECEDENCE)
+@RequiredArgsConstructor
 @RestControllerAdvice
 public class GlobalExceptionAdvice {
+
+    private final MessageSource ms;
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ErrorResponse MemberNotFoundException(MethodArgumentNotValidException exception){
+        BindingResult bindingResult = exception.getBindingResult();
+        String[] codes = bindingResult.getAllErrors().get(0).getCodes();
+        for (String code : codes) {
+            System.out.println("code = " + code);
+        }
+
+        String code = codes[1];
+        return new ErrorResponse(ms.getMessage(code,null,null));
+    }
+
+    @Getter
+    public class ErrorResponse {
+        String message;
+        public ErrorResponse(String message) {
+            this.message = message;
+        }
+    }
+
+    @ExceptionHandler(BaseException.class)
+    public BaseResponse<BaseResponseStatus> baseException(BaseException e) {
+        log.error("Handle CommonException: {}", e.getMessage());
+        return new BaseResponse<>(e.getStatus());
+    }
+
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    @ExceptionHandler({ UnsupportedEncodingException.class, NoSuchAlgorithmException.class, InvalidKeyException.class, JsonProcessingException.class} )
+    public BaseResponse<BaseResponseStatus> smsHandleException(Exception e) {
+        //log.error("Handle All Exception: {}", e.getMessage());
+        return new BaseResponse<>(BaseResponseStatus.SMS_API_ERROR);
+    }
 
     @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
     @ExceptionHandler(Exception.class)
